Data.ByteString.Search.Substitution

Description

Class for values to be substituted into strict and lazy ByteStrings by the replace functions defined in this package.

Documentation

class Substitution a where

Type class of meaningful substitutions for replace functions on ByteStrings. Instances for strict and lazy ByteStrings are provided here.

Methods

substitution :: a -> [ByteString] -> [ByteString]

substitution transforms a value to a substitution function.

prependCycle :: a -> ByteString -> ByteString

prependCycle sub lazyBS shall prepend infinitely many copies of sub to lazyBS without entering an infinite loop in case of an empty sub, so e.g.

  prependCycle "" "ab" == "ab"

shall (quickly) evaluate to True. For non-empty sub, the cycle shall be constructed efficiently.